Update oc2slpf-v1.0-wip.md
A new commit. It seems that accidentally i had deleted part of 2.1.2-1. Its "fixed" now.

The following are comments from the previous commit and pull request.
-We need to be consistent in the way we write. For example capitalization vs small letter in specific key words (Specifiers, Options, Arguments, Section, Version, OpenC2 Language Specification).

-Options vs. Command Arguments. They are used interchangeably. This may confuse the reader. In the Abstract we have the same "issue" with the word options. I sugget to change it to arguments.

-"In the context of" a SLPF vs an SLPF vs Stateless Packet Filtering.
(The reader will get a bit confused because you have to understand that the first refers to "a stateless packet filter", the second "an SLPF" (we define an entity - device - a defence system))

-figure 1.1 needs to be centered.